Added caching to user session manager

+++ src/protected/components/UserSessionManager.php
@@ -36,6 +36,9 @@ class UserSessionManager extends CComponent {
 		
 		$session = UserSession::model()->find('token=:token', array(':token'=>$token));
 		if($session != null) {
+			$cacheKey = self::buildSessionCacheKey($session->user_id, $token);
+			CacheManager::getInstance()->delete($cacheKey);
+			
 			$session->delete();
 		}
 	}
@@ -46,6 +49,10 @@ class UserSessionManager extends CComponent {
 		if($token != null) {
 			self::destroySession($token);
 		}
+	}
+	
+	private static function buildSessionCacheKey($user_id, $token) {
+		return "session:".$user_id.$token;
 	}
 	
 	/**
@@ -68,39 +75,43 @@ class UserSessionManager extends CComponent {
 			return null;
 		}
 		
-		// Try to find session with matching user_id and token
-		$session = UserSession::model()->find('user_id=:user_id AND token=:token', array(':user_id'=>$user_id, ':token'=>$token));
-		
-		if($session == null) {
-			// No matching session in DB. Invalid.
-			self::clearAllSessionCookies();
-			return null;
-		} else if(CDateTimeParser::parse($session->expires,'yyyy-MM-dd hh:mm:ss') < time()) {
-			// Session has expired and therefor invalid. Clean it up.
-			$session->delete();
-			self::clearAllSessionCookies();
-			return null;
+		// Search the cache for session nonce before hitting the DB
+		$cacheKey = self::buildSessionCacheKey($user_id, $token);
+		if(!($local_nonce = CacheManager::getInstance()->get($cacheKey))) {
+			// Try to find session with matching user_id and token
+			$session = UserSession::model()->find('user_id=:user_id AND token=:token', array(':user_id'=>$user_id, ':token'=>$token));
+			
+			if($session == null) {
+				// No matching session in DB. Invalid.
+				self::clearAllSessionCookies();
+				return null;
+			} else if(CDateTimeParser::parse($session->expires,'yyyy-MM-dd hh:mm:ss') < time()) {
+				// Session has expired and therefor invalid. Clean it up.
+				$session->delete();
+				self::clearAllSessionCookies();
+				return null;
+			}
+			$local_nonce = $session->nonce;
 		}
 		
 		// Validate the nonce
-		$nonce = self::$dirty ? self::$last_data['nonce'] : self::getCookie(self::COOKIE_SESSION_NONCE);
+		$read_nonce = self::$dirty ? self::$last_data['nonce'] : self::getCookie(self::COOKIE_SESSION_NONCE);
 		
-		if($nonce == null) {
+		if($read_nonce == null) {
 			// No nonce? No go
 			self::destroySession($token);
 			return null;
-		} else if ($nonce != $session->nonce) {
+		} else if ($read_nonce != $local_nonce) {
 			// Nonce doesn't match expected indicates clear case of tampering (specifically, copied cookies on multiple devices)
 			self::destroySession($token);
 			return null;
 		}
 		
 		// Everything looks good! Update the nonce and return the user ID
-		$new_nonce = self::generateRandomToken($user_id);
-		self::setCookie(self::COOKIE_SESSION_NONCE, $new_nonce);
-		self::setDirtyData($user_id, $token, $new_nonce);
-		$session->nonce = $new_nonce;
-		$session->save();
+		$new_nonce = self::updateSessionNonce($user_id, $token);
+		
+		// Save the new_nonce in the cache for an hour.
+		CacheManager::getInstance()->set($cacheKey, $new_nonce, 60 * 60);
 		
 		return $user_id;
 	}
@@ -113,6 +124,24 @@ class UserSessionManager extends CComponent {
 		if($user_id == null) {
 			return false;
 		}
+	}
+	
+	/**
+	 * This method generates a new nonce then saves it in the user's cookies and the DB
+	 * @param unknown $user_id
+	 * @param unknown $token
+	 */
+	private static function updateSessionNonce($user_id, $token) {
+		$new_nonce = self::generateRandomToken($user_id);
+		self::setCookie(self::COOKIE_SESSION_NONCE, $new_nonce);
+		self::setDirtyData($user_id, $token, $new_nonce);
+		
+		$session = UserSession::model()->find('user_id=:user_id AND token=:token', array(':user_id'=>$user_id, ':token'=>$token));
+		$session->nonce = $new_nonce;
+		$session->expires = new CDbExpression('NOW() + INTERVAL 30 DAY');
+		$session->save();
+		
+		return $new_nonce;
 	}
 	
 	/**
